virNetworkDHCPLeaseTimeDef: Make expiry unsigned long long
The width of `unsigned long` differs on 32 bit and 64 bit architectures.
There is no compelling reason why the maximum DHCP lease time should
depend on the architecture.

+++ src/conf/network_conf.h
@@ -105,7 +105,7 @@ VIR_ENUM_DECL(virNetworkForwardDriverName);
 
 typedef struct _virNetworkDHCPLeaseTimeDef virNetworkDHCPLeaseTimeDef;
 struct _virNetworkDHCPLeaseTimeDef {
-    unsigned long expiry;
+    unsigned long long expiry;
     virNetworkDHCPLeaseTimeUnitType unit;
 };
